move test to the right location
